api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openai
python-dotenv
azure-cosmos 
azure-identity
azure-monitor-opentelemetry